Problem: webhook module is hogging the webserver"
Solution: Externalize the webserver.

This commit externalizes the webserver and refactors the webook_target
package to instead abstract general interaction with github.  The gin
routing package is used as the web router, and an abstraction layer on
top of this is built to register handlers for arbitrary routes and
verbs.